refactor: Move user settings related endpoints into SettingController
